This commit was manufactured by cvs2svn to create tag 'r242'.

+++ Modules/getpath.c
@@ -381,7 +381,7 @@ calculate_path(void)
     NSModule pythonModule;
 #endif
 #ifdef __APPLE__
-    uint32_t nsexeclength = MAXPATHLEN;
+    unsigned long nsexeclength = MAXPATHLEN;
 #endif
 
 	/* If there is no slash in the argv0 path, then we have to
